Improved DriveStraight commands
They now include periods for acceleration and deceleration. Also, the reason why DriveStraightByDistance went too far is because the maximum velocity does not equal the maximum power. You tried to find the time with the power, not the velocity. vmax is the maximum encoder rate when the robot is moving at maximum power. It has been 2 for us in the past.